requirements.txt
assimp_py==1.0.7
glfw==2.5.4
imgui==2.0.0
numpy==1.23.5
open3d==0.16.0
Pillow==9.3.0
Pillow==10.0.0
plyfile==0.7.4
PyGLM==2.6.0
PyOpenGL==3.1.6
pyrr==0.10.3
